testing_1/requirements.txt
celery
celery[redis]
PyPDF2
chromadb
langchain_community
langchain
sentence-transformers
python-multipart
sqlalchemy
python-docx
pymongo